Make on linux Add CLI docs from the cli output
