+++ docs/cli-experiments.rst
@@ -4,4 +4,4 @@ Experiments
 
 ...To see the list of included experiment commands type the following at a command prompt::
 
-        idmtools simulation --help
+.. command-output:: idmtools experiment --help
