+++ docs/cli-simulations.rst
@@ -4,4 +4,4 @@ Simulations
 
 ...To see the list of included simulation commands following at a command prompt::
 
-        idmtools simulation --help
+.. command-output:: idmtools simulation --help
