+++ docs/cli-templates.rst
@@ -2,7 +2,6 @@
 Templates
 =========
 
-You can use the cookiecutter templates included with |IT_s| to get started with python projects and |IT_s|. These templates provide a logical, reasonably standardized, but flexible project structure for doing and sharing data science work. To see the list of included cookiecutter templates type the following at a command prompt::
+You can use the cookiecutter templates included with |IT_s| to get started with python projects and |IT_s|. These templates provide a logical, reasonably standardized, but flexible project structure for doing and sharing data science work. To see the list of included cookiecutter templates type the following at a command prompt.
 
-        idmtools init --help
-
+.. command-output:: idmtools init --help
