+++ docs/cli_index.rst
@@ -12,12 +12,7 @@ CLI reference
    cli-platforms
    cli-troubleshoot
    
-|IDM_s| includes a command-line interface (CLI) with options and commands to assist with getting started, managing and monitoring, and troubleshooting simulations and experiements. After you've installed |IDM_s| you can view the available options and commands by typing the following at a command prompt::
+|IDM_s| includes a command-line interface (CLI) with options and commands to assist with getting started, managing and monitoring, and troubleshooting simulations and experiements. After you've installed |IDM_s| you can view the available options and commands by typing the following at a command prompt
 
-.. click:: idmtools_cli.cli:cli
-  :prog: idmtools
-  :show-nested:
+.. command-output:: idmtools --help
 
-.. click:: idmtools_cli.cli.init:init
-  :prog: idmtools
-  :show-nested:
