+++ docs/requirements.txt
@@ -3,4 +3,4 @@ sphinxcontrib-napoleon
 beautifulsoup4
 sphinx
 plantweb
-sphinx-click
+sphinxcontrib-programoutput
