+++ idmtools_cli/idmtools_cli/cli/experiment.py
@@ -3,7 +3,7 @@ import click
 
 from idmtools_cli.iplatform_cli import IPlatformCLI
 
-from idmtools_cli.cli import cli, pass_platform_cli
+from idmtools_cli.cli.entrypoint import pass_platform_cli, cli
 from idmtools_cli.cli.utils import tags_help, get_platform_from_config_or_name, supported_platforms
