+++ idmtools_cli/idmtools_cli/cli/simulation.py
@@ -2,8 +2,7 @@ from typing import Optional, Tuple, List
 import click
 
 from idmtools_cli.iplatform_cli import IPlatformCLI
-from idmtools_cli.cli import cli
-from idmtools_cli.cli.experiment import pass_platform_cli
+from idmtools_cli.cli.entrypoint import pass_platform_cli, cli
 from idmtools_cli.cli.utils import tags_help, get_platform_from_config_or_name, supported_platforms
