+++ idmtools_platform_local/idmtools_platform_local/cli/local.py
@@ -5,7 +5,7 @@ import docker
 import stringcase as stringcase
 from colorama import Fore
 
-from idmtools_cli.cli import cli
+from idmtools_cli.cli.entrypoint import cli
 from idmtools_platform_local.cli.utils import get_service_info
 from idmtools_platform_local.infrastructure.docker_io import DockerIO
 from idmtools_platform_local.infrastructure.service_manager import DockerServiceManager
