+++ idmtools_platform_local/idmtools_platform_local/internals/tasks/docker_run.py
@@ -88,23 +88,26 @@ class DockerBaseTask(BaseTask):
             return -1
 
 
-class DockerRunTask(GenericActor, DockerBaseTask):
-    class Meta:
-        store_results = False
-        max_retries = 0
-        queue_name = "cpu"
+try:
+    class DockerRunTask(GenericActor, DockerBaseTask):
+        class Meta:
+            store_results = False
+            max_retries = 0
+            queue_name = "cpu"
 
-    def perform(self, command: str, experiment_uuid: str, simulation_uuid: str, container_config: dict) -> Status:
-        return self.docker_perform(command, experiment_uuid, simulation_uuid, container_config)
+        def perform(self, command: str, experiment_uuid: str, simulation_uuid: str, container_config: dict) -> Status:
+            return self.docker_perform(command, experiment_uuid, simulation_uuid, container_config)
 
 
-# it would be great we could just derive from RunTask and change the meta but that doesn't seem to work with
-# GenericActors for some reason. Using BaseTask and these few lines of redundant code are our compromise
-class GPURunTask(GenericActor, DockerBaseTask):
-    class Meta:
-        store_results = False
-        max_retries = 0
-        queue_name = "gpu"
+    # it would be great we could just derive from RunTask and change the meta but that doesn't seem to work with
+    # GenericActors for some reason. Using BaseTask and these few lines of redundant code are our compromise
+    class GPURunTask(GenericActor, DockerBaseTask):
+        class Meta:
+            store_results = False
+            max_retries = 0
+            queue_name = "gpu"
 
-    def perform(self, command: str, experiment_uuid: str, simulation_uuid: str, container_config: dict) -> Status:
-        return self.docker_perform(command, experiment_uuid, simulation_uuid, container_config)
+        def perform(self, command: str, experiment_uuid: str, simulation_uuid: str, container_config: dict) -> Status:
+            return self.docker_perform(command, experiment_uuid, simulation_uuid, container_config)
+except TypeError:  # Doc mocks not working
+    pass
