+++ idmtools_test/idmtools_test/utils/cli.py
@@ -23,7 +23,7 @@ def invoke_command(*args, start_command=None):
     if start_command is None:
         start_command = []
     from idmtools_cli.main import start
-    from idmtools_cli.cli import cli
+    from idmtools_cli.cli.entrypoint import cli
     start()
     runner = CliRunner()
     final_command = start_command + list(args) if len(args) else start_command
